Create Condensed Arrays
Write a program to read an array of integers and condense them by summing adjacent couples of elements until a single integer is obtained. 
For example, if we have 3 elements {2, 10, 3}, we sum the first two and the second two elements and obtain {2+10, 10+3} = {12, 13}.
Then we sum again all adjacent elements and obtain {12+13} = {25} and print.